requirements.txt
flask~=3.0
flask-cors~=4.0
sqlalchemy~=2.0
python-dotenv~=1.0
openai~=1.102
